+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 
+## [0.0.3] - 2021/07/11. 
+* Renamed All Class
+* Finished Animation Sprite 
+* Finished Text Sprite
+* Finished ImageSprite Sprite
+* Finished Support TexturePacker Plist And Json Sprite
+* Finished Base Demo
+
 ## [0.0.2] - 2021/07/07. 
 * GameScene
 * GameRenderBox 
-* FpsSprite MapSprite PlayerSprite MonsterSprite 
+* Assets Loader
 
 ## [0.0.1] - 2021/07/06. 
-* GameLoop 
-* Game 
-* Sprite 
+* Game Loop 
+* Game Widget
+* Base Sprite